Removed some coupling
