+++ package.json
@@ -3,12 +3,13 @@
   "version": "5.3.2",
   "description": "Create your next immutable state by mutating the current one",
   "main": "dist/immer.js",
+  "module": "dist/immer.module.js",
   "umd:main": "dist/immer.umd.js",
   "unpkg": "dist/immer.umd.js",
   "jsdelivr": "dist/immer.umd.js",
-  "module": "dist/immer.module.js",
   "jsnext:main": "dist/immer.module.js",
   "react-native": "dist/immer.module.js",
+  "source": "src/index.ts",
   "types": "./dist/index.d.ts",
   "scripts": {
     "test": "jest && yarn-or-npm test:build && yarn-or-npm test:flow",
